[item] typo fix

+++ models/item.py
@@ -11,7 +11,7 @@ class ItemModel(db.Model):
     name: Mapped[str] = mapped_column(db.String(255), nullable=False)
     price: Mapped[str] = mapped_column(db.Float(), nullable=False)
     part_number: Mapped[str] = mapped_column(db.String(32), nullable=False, unique=True)
-    ean: Mapped[int] = mapped_column(db.Integer(13))
+    ean: Mapped[int] = mapped_column(db.Integer())
     category: Mapped[str] = mapped_column(db.String(32), nullable=False)
     specs: Mapped[dict] = mapped_column(JSON, nullable=True)
     stocks: Mapped[int] = mapped_column(db.Integer(), nullable=False, default=0)
